requirements.txt
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
Pillow==10.0.0
pandas==2.0.3